Core/pyproject.toml
[tool.poetry]
name = "crm-api"
version = "0.1.0"
description = "CRM API"
authors = ["tienpv <tienpv@saisystem.vn>"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "0.18.2"
gunicorn = "20.1.0"
fastapi = "0.78.0"
PyJWT = "2.1.0"
python-box = "6.0.2"
boto3 = "1.24.22"
simplejson = "3.17.6"
Jinja2 = "3.1.2"
jsonschema = "4.6.1"
pytz = "2022.1"
reportlab = "3.6.2"
Pillow = "8.4.0"
qrcode = "7.3.1"
PyPDF2 = "1.26.0"
pyotp = "2.6.0"
PyMuPDF = "1.20.1"
SQLAlchemy = "1.4.37"
mysqlclient = "2.1.1"
python-dotenv = "0.20.0"
pydantic = "1.9.1"
requests = "2.28.0"
tomlkit = "0.11.0"
dependency-injector = "4.40.0"
watchtower = "3.0.0"
python-json-logger = "2.0.4"
celery = "5.2.7"
eventlet = "0.33.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
